backend/requirements.txt
asgiref==3.8.1
cffi==1.16.0
cryptography==42.0.8
Django==5.0.6
django-cors-headers==4.4.0
django-health-check==3.18.3
django-rest-knox==4.2.0
djangorestframework==3.15.2
gunicorn==22.0.0
packaging==24.1
psycopg2-binary==2.9.9
pycparser==2.22
sqlparse==0.5.0
uuid==1.30
whitenoise==6.7.0
